feat: handleMouseUp 분리
